syntax and comment changes.
